A quick demonstration of using BSP drivers to use the built in LCD screen on the 769 Discovery.
A simple demonstration of using the BSP drivers to use the built in LCD. It involves creating the initial project in CUBEMX and then copying the BSP drivers to the project folder and also the utilities and fonts folders to the project. Updating the cubemxide project properties to include the newly added files and adjusting the code with the proper initializations. I hope this helps someone.